ld carry bug and xor bug fixed

+++ src/vm/instructions/inst08_xor.c
@@ -6,7 +6,7 @@
 /*   By: mnishimo <mnishimo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/29 20:44:18 by mnishimo          #+#    #+#             */
-/*   Updated: 2019/06/24 12:20:05 by mnishimo         ###   ########.fr       */
+/*   Updated: 2019/06/24 13:00:29 by mnishimo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,10 +21,10 @@ void	inst_xor(t_game *game, t_process *caller, t_inst *inst)
 		return ;
 	if (!(i = get_arg(caller, game->memdump, &(inst->args[0]), 1)))
 		return ;
-	res = (int)*i;
+	res = *i;
 	if (!(i = get_arg(caller, game->memdump, &(inst->args[1]), 1)))
 		return ;
-	res = res ^ (int)*i;
+	res = res ^ *i;
 	caller->carry = (!res) ? 1 : 0;
 	ft_memcpy(get_arg(caller, game->memdump, &(inst->args[2]), 0), &res, REG_SIZE);
 	if (game->deb_state & 32)
